Make sure the database is using 64-bit ints

+++ server/src/database/Database.kt
@@ -3,8 +3,8 @@ package leif.database
 interface Database {
     fun selectOne(query: String, params: List<Any?> = emptyList()): Map<String, Any?>?
     fun select(query: String, params: List<Any?> = emptyList()): List<Map<String, Any?>>
-    fun insert(query: String, params: List<Any?> = emptyList()): List<Int>
-    fun update(query: String, params: List<Any?> = emptyList()): Int
+    fun insert(query: String, params: List<Any?> = emptyList()): List<Long>
+    fun update(query: String, params: List<Any?> = emptyList()): Long
     fun delete(query: String, params: List<Any?> = emptyList()): Boolean
     fun statement(query: String, params: List<Any?> = emptyList()): Unit
     fun <R> transaction(fn: (Database) -> R): R
